Operators Created
Each elementary operator has been added:
    +
    -
    *
    /

Each operator has a check, bParam, to see if the
second provided variable "b" has been given. If it
hasn't, then like a digital calculator it will
calculate the result based on using the variable
"a" for both "a" and "b" in the calculation.