Altered Operators, add class red
The operators "x" and "-" were originally using
the ASCII codes for the mathematical Operators
and not what's on anyones keyboard. This has been
changed to now make it easier to read for anyone
looking at the hmtl doc.

Class "red" was added to the button with "ON/C" to it.
It will be used to change that sole button red.