Renamed btnIDs + numPow
For the container, for readability, I changed
btn_numsPow > cntnr_numsPow so that it's easier to distinguish
that it's a container and not a button.

All the button ID's which started with "btn_" and are associated
with either numbers or operators have been changed to better
reflect what they are when targetting them within JS for DOM
manipulation.

btn_# > btnNum_#
btn_x > btnOp_x

Even though the decimal point for "btn_." technically isn't a number,
it's lumped into the same category since they'll be appended to
the final string.