Button Programming-Nums
I started programming the JS side of the calculator, in this aspect
I'm focusing on the numbers themselves. The variables are currently
what I think I'll need to use but may be dropped over time.

The next portion I plan to do is clearing the screen, and making
sure that only one decimal can exist for the currently displayed
number.