num AnonFunc > funct & Keyboard Support
=== numButtons Anonymous Function  >  numBtn() ===

The anonymous function with this variable has been converted
into a function: numBtn(). In addition to this, the function
has been updated to support both interaction via the users
mouse, as well as their keyboard.

=== Keyboard Support ===

The beginnings of keyboard support for working the calculator
have been added!

=== New Function: keyOpConversion() ===

Converts the "/" and "*" strings on the keyboard to their
respective operatinal counterparts.